- generate packages - generate sources - creates different Linux_for_Tegra_XXX folders for different boards - prepared 36.2 but doesn't compile - List of working "L4T version : SOM / boards" tuples :    - 35.1 : Xavier NX 16GB  / Xavier devkit    - 35.3.1 : Xavier NX 16GB  / Xavier devkit    - 35.3.1 : AGX Orin  / Auvidea X230D    - 35.4.1 : Xavier NX 16GB  / Xavier devkit    - 35.5.0 : Xavier NX 16GB  / Xavier devkit

+++ l4t_clean_diff.sh
@@ -1,18 +1,24 @@
 . environment $@
 
-pushd $L4T_VERSION/Linux_for_Tegra/
+if [[ ! -d $L4T_VERSION/${LINUX_FOR_TEGRA_DIR} ]]
+then
+   echo "Error : $L4T_VERSION/${LINUX_FOR_TEGRA_DIR} folder doesn't exist"
+   exit
+fi
+
+pushd $L4T_VERSION/$LINUX_FOR_TEGRA_DIR/
 git ls-files -m | xargs git rm --cached
 git stash
 git clean -f
 popd
 
-pushd $L4T_VERSION/Linux_for_Tegra/source/public/kernel/
+pushd $L4T_VERSION/$LINUX_FOR_TEGRA_DIR/source/public/kernel/
 git ls-files -m | xargs git rm --cached
 git stash
 git clean -f
 popd
 
-pushd $L4T_VERSION/Linux_for_Tegra/source/public/hardware/
+pushd $L4T_VERSION/$LINUX_FOR_TEGRA_DIR/source/public/hardware/
 git ls-files -m | xargs git rm --cached
 git stash
 git clean -f
